chore: externalize the prev/next buttons to it's component
